[Refactoring] Move Authorizator from mailbox-store to mailbox-api
